Fixed Auth error, now there's another
